requirements.txt
fastapi
uvicorn[standard]
openai
aiohttp
pillow
python-dotenv
pytest
openai-whisper
requests